+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Core/Arbitrary.hs
@@ -535,6 +535,8 @@ deriving instance (Era era, Arbitrary (PParamsHKD StrictMaybe era)) => Arbitrary
 
 deriving instance Arbitrary CoinPerByte
 
+deriving newtype instance Arbitrary MaxPledgeLeverage
+
 ------------------------------------------------------------------------------------------
 -- Cardano.Ledger.DRep -------------------------------------------------------------------
 ------------------------------------------------------------------------------------------